requirements.txt
google-adk>=0.1.0
google-generativeai>=0.8.0
Pillow>=10.0.0
asyncio-python>=3.9.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0